+++ lib/models/rating.dart
@@ -4,13 +4,13 @@ class Rating {
   String id;
   String userId;
   String portfolioId;
-  int stars;
+  double stars;
 
   Rating(this.id, this.userId, this.portfolioId, this.stars);
 
   factory Rating.fromFirestore(DocumentSnapshot doc) {
     Map data = doc.data;
-    return Rating(doc.documentID, data['userId'], data['portfolioId'], data['stars']);
+    return Rating(doc.documentID, data['userId'], data['portfolioId'], data['stars'].toDouble());
   }
 
   Map<String, dynamic> toMap() {
